# contracts/openapi/core-api.yaml
openapi: "3.1.0"
info:
  title: HealthOS Core API
  version: "1.0.0"
  description: |
    REST API cho Core Backend Service của HealthOS.
    Base URL: http://localhost:8000
    Tất cả endpoint đặt dưới /v1/
servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://api.healthos.example.com
    description: Production
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: NOT_FOUND
            message:
              type: string
              example: Resource not found.
            details:
              type: object
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total:
          type: integer
          example: 100
...
    MealResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        status:
          type: string
          enum: [pending, processing, analyzed, failed]
        nutrition_result:
          $ref: "#/components/schemas/NutritionResult"
        logged_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
    NutritionResult:
      type: object
      properties:
        calories:
          type: number
          format: float
        protein_g:
          type: number
          format: float
        carbs_g:
          type: number
          format: float
        fat_g:
          type: number
          format: float
        confidence:
          type: number
          format: float
          minimum: 0
          maximum: 1
...
paths:
...
  /v1/meals:
    get:
      tags: [Meals]
      summary: List meal logs
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: per_page
          in: query
          schema: { type: integer, default: 20 }
        - name: date_from
          in: query
          schema: { type: string, format: date }
        - name: date_to
          in: query
          schema: { type: string, format: date }
      responses:
        "200":
          description: Paginated meal list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/MealResponse"
                  meta:
                    $ref: "#/components/schemas/PaginationMeta"
        "400":
          description: Invalid query params
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      tags: [Meals]
      summary: Create a meal log (with optional image upload)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                notes:
                  type: string
                logged_at:
                  type: string
                  format: date-time
          multipart/form-data:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                notes:
                  type: string
                logged_at:
                  type: string
                  format: date-time
                image:
                  type: string
                  format: binary
      responses:
        "201":
          description: Meal created, analysis enqueued
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/MealResponse"
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/meals/{meal_id}:
    get:
      tags: [Meals]
      summary: Get single meal with nutrition result
      parameters:
        - name: meal_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Meal detail
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/MealResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /v1/notifications:
    get:
      tags: [Notifications]
      summary: List notifications (cursor-paginated, newest first)
      parameters:
        - { name: cursor, in: query, schema: { type: string } }
        - { name: per_page, in: query, schema: { type: integer, default: 20, maximum: 50 } }
        - { name: only_unread, in: query, schema: { type: boolean, default: false } }
      responses:
        "200":
          description: Notification list with pagination meta
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { type: object } }
                  meta:
                    type: object
                    properties:
                      next_cursor: { type: string, nullable: true }
                      has_more: { type: boolean }
                      per_page: { type: integer }
        "401": { $ref: "#/components/schemas/ErrorResponse" }
...
  /v1/notifications/{notification_id}/read:
    post:
      tags: [Notifications]
      summary: Mark a single notification as read (idempotent)
      parameters:
        - { name: notification_id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200": { description: Notification envelope (with read_at populated) }
        "404": { $ref: "#/components/schemas/ErrorResponse" }
...